agricultural-produce-classifier/requirements.txt
tensorflow==2.10.0
keras==2.10.0
opencv-python==4.6.0.66
numpy==1.23.3
Pillow==9.2.0
matplotlib==3.5.3
scikit-learn==1.1.2
pandas==1.4.4
jupyter==1.0.0
notebook==6.4.12
